Add cumulative card event (#74)
* Update `jokers_of_neon_lib` dependency branch to `event-accumulative-card` in Scarb.toml files and change `game_id` type from `u32` to `u64` in multiple Cairo modules for improved compatibility and consistency.

* Fix typo in `jokers_of_neon_lib` dependency branch in Scarb.toml by removing trailing space for consistency.

* Update `jokers_of_neon_lib` dependency branch to `develop` in Scarb.toml files for consistency across projects.

+++ mods/jokers_of_neon_classic/src/specials/play/burning_rewards.cairo
@@ -43,7 +43,7 @@ pub mod special_burning_rewards {
 
     #[abi(embed_v0)]
     impl BurningRewardsInfo of ICardInfo<ContractState> {
-        fn values(self: @ContractState, game_id: u32) -> (i32, i32, i32) {
+        fn values(self: @ContractState, game_id: u64) -> (i32, i32, i32) {
             let mut world = self.world(@"jokers_of_neon_classic");
             let cumulative: Cumulative = world.read_model((game_id, BURNING_REWARDS_KEY));
             (cumulative.value, 0, 0)
