+++ mods/jokers_of_neon_classic/src/specials/play/high_roller.cairo
@@ -62,7 +62,7 @@ pub mod special_high_roller {
 
     #[abi(embed_v0)]
     impl HighRollerInfo of ICardInfo<ContractState> {
-        fn values(self: @ContractState, game_id: u32) -> (i32, i32, i32) {
+        fn values(self: @ContractState, game_id: u64) -> (i32, i32, i32) {
             let mut world = self.world(@"jokers_of_neon_classic");
             let cumulative: Cumulative = world.read_model((game_id, HIGH_ROLLER_KEY));
             (0, cumulative.value, 0)
